requirements.txt
python-dotenv
fastapi
uvicorn
sqlmodel
alembic
psycopg2
pytest
httpx
